Create template workout with timer
I created a template pug file with the intention of progressive enhancement
to give users who do not have access to JavaScript a list of the exercises
included in a routine.

For users with JavaScript, I imported a timer from a [previous project](github.com/mcksheridan/timer)
to use here.

I passed in some sample data on the server to use in the template.